requirements.txt
Django==3.1.6
django-crispy-forms==1.11.0
django-debug-toolbar==3.2
django-debug-toolbar==3.2
Pillow==8.1.0


appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.3.1
asn1crypto==0.24.0
astroid==2.4.2
beautifulsoup4==4.6.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
configobj==5.0.6
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distlib==0.3.1

filelock==3.0.12
httplib2==0.9.2
idna==2.6
importlib-metadata==3.4.0
importlib-resources==5.1.0
isort==5.7.0
keyring==10.6.0
keyrings.alt==3.0
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
nemo-emblems==4.4.0
netifaces==0.10.4
onboard==1.4.1
PAM==0.4.2
pbr==5.5.1
pexpect==4.2.1
protobuf==3.0.0
psutil==5.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygobject==3.26.1
PyICU==1.9.8
pyinotify==0.9.6
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparted==3.11.1
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.5
python-debian==0.1.32
python-xapp==1.8.1
python-xlib==0.20
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
sessioninstaller==0.0.0
setproctitle==1.1.10
six==1.15.0
sqlparse==0.4.1
stevedore==3.3.0
system-service==0.3
systemd-python==234
tinycss==0.4
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.0.22
urllib3==1.22
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wrapt==1.12.1
xkit==0.0.0
zipp==3.4.0
